Version 1 of V27 model for OpenFASTv1.0.0
Version 1 of V27 model, tested with OpenFASTv1.0.0.  Based on the FAST version 7 model of the V27 from Chris Kelley, 10-16-2018.  The AeroDyn blade input was interpolated based on the FASTv7 model with an additional node at max chord. The number of nodes was increased to 27 to place a node at max chord (2m from rotor center via Resor 2014 report SAND2014-19136), with chord measurements and twist from report SAND2018-11893. A previous version of this OF planform was incorrect, this version has been corrected to more closely match the FASTv7 model, change in power was 1.2% at 6.5mps and 1.1% at 8.3mps due to these corrections and the new max chord node.